# apim-lab/openapi.json
{"openapi":"3.1.0","info":{"title":"Calculator API","version":"1.0"},"paths":{"/":{"get":{"summary":"Home","operationId":"home__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/add":{"get":{"summary":"Add","operationId":"add_add_get","parameters":[{"name":"a","in":"query","required":true,"schema":{"type":"number","title":"A"}},{"name":"b","in":"query","required":true,"schema":{"type":"number","title":"B"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/sub":{"get":{"summary":"Sub","operationId":"sub_sub_get","parameters":[{"name":"a","in":"query","required":true,"schema":{"type":"number","title":"A"}},{"name":"b","in":"query","required":true,"schema":{"type":"number","title":"B"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mul":{"get":{"summary":"Mul","operationId":"mul_mul_get","parameters":[{"name":"a","in":"query","required":true,"schema":{"type":"number","title":"A"}},{"name":"b","in":"query","required":true,"schema":{"type":"number","title":"B"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/div":{"get":{"summary":"Div","operationId":"div_div_get","parameters":[{"name":"a","in":"query","required":true,"schema":{"type":"number","title":"A"}},{"name":"b","in":"query","required":true,"schema":{"type":"number","title":"B"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}